get taxes from planets

+++ Domain/Factions/IFactionTaxService.cs
@@ -2,7 +2,7 @@ namespace SkyHorizont.Domain.Factions
 {
     public interface IFactionTaxService
     {
-        void TaxPlanet(Guid planetId, int percentage);
+        void TaxPlanet(Guid planetId, double percentage);
         void DistributeLoot(Guid leaderCharacterId, int totalLoot, IEnumerable<Guid> subCharacterIds);
     }
 }
